Allow client owner to regenerate secret

+++ iam-login-service/src/main/java/it/infn/mw/iam/util/IamClientSecretEncoder.java
@@ -24,7 +24,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
  */
 public class IamClientSecretEncoder implements PasswordEncoder {
 
-  final short DEFAULT_ROUND = 12;
+  final int DEFAULT_ROUND = 12;
 
   BCryptPasswordEncoder bcryptEncoder = new BCryptPasswordEncoder(DEFAULT_ROUND);
